ai-server/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pillow==10.0.0
pillow==10.0.0
scikit-learn>=1.3.0
numpy>=1.24.3
gunicorn==21.2.0
